+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 ## mini.files
 
 - FEATURE: prefer using 'mini.icons' as icon provider.
+- FEATURE: make data for `MiniFilesActionDelete` contain `to` field in case of not permanent delete.
 - BREAKING: update how confirmation lines are computed:
     - Show create actions in the group directory where text manipulation took place. This matters during creating nested entries and is usually a more intuitive representation.
     - For delete show its type after the file name ("permanently" or "to trash") as an additional visual indication of delete type.